# pkg/models/error_result.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	strfmt "github.com/go-openapi/strfmt"

	"github.com/go-openapi/errors"
	"github.com/go-openapi/swag"
	"github.com/go-openapi/validate"
)

// ErrorResult error result
// swagger:model errorResult
type ErrorResult struct {

	// код ошибки
	// Required: true
	Code string `json:"code"`

	// описаниие ошибки
	// Required: true
	Message string `json:"message"`
}

// Validate validates this error result
func (m *ErrorResult) Validate(formats strfmt.Registry) error {
	var res []error

	if err := m.validateCode(formats); err != nil {
		res = append(res, err)
	}

	if err := m.validateMessage(formats); err != nil {
		res = append(res, err)
	}

	if len(res) > 0 {
		return errors.CompositeValidationError(res...)
	}
	return nil
}

func (m *ErrorResult) validateCode(formats strfmt.Registry) error {

	if err := validate.RequiredString("code", "body", string(m.Code)); err != nil {
		return err
	}

	return nil
}

func (m *ErrorResult) validateMessage(formats strfmt.Registry) error {

	if err := validate.RequiredString("message", "body", string(m.Message)); err != nil {
		return err
	}

	return nil
}

// MarshalBinary interface implementation
func (m *ErrorResult) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *ErrorResult) UnmarshalBinary(b []byte) error {
	var res ErrorResult
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}