# pkg/models/get_universities_o_k_body_universities_items.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	strfmt "github.com/go-openapi/strfmt"

	"github.com/go-openapi/errors"
	"github.com/go-openapi/swag"
	"github.com/go-openapi/validate"
)

// GetUniversitiesOKBodyUniversitiesItems get universities o k body universities items
// swagger:model getUniversitiesOKBodyUniversitiesItems
type GetUniversitiesOKBodyUniversitiesItems struct {

	// идентификтора университета
	// Required: true
	ID int64 `json:"id"`

	// название университета
	// Required: true
	Name string `json:"name"`
}

// Validate validates this get universities o k body universities items
func (m *GetUniversitiesOKBodyUniversitiesItems) Validate(formats strfmt.Registry) error {
	var res []error

	if err := m.validateID(formats); err != nil {
		res = append(res, err)
	}

	if err := m.validateName(formats); err != nil {
		res = append(res, err)
	}

	if len(res) > 0 {
		return errors.CompositeValidationError(res...)
	}
	return nil
}

func (m *GetUniversitiesOKBodyUniversitiesItems) validateID(formats strfmt.Registry) error {

	if err := validate.Required("id", "body", int64(m.ID)); err != nil {
		return err
	}

	return nil
}

func (m *GetUniversitiesOKBodyUniversitiesItems) validateName(formats strfmt.Registry) error {

	if err := validate.RequiredString("name", "body", string(m.Name)); err != nil {
		return err
	}

	return nil
}

// MarshalBinary interface implementation
func (m *GetUniversitiesOKBodyUniversitiesItems) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *GetUniversitiesOKBodyUniversitiesItems) UnmarshalBinary(b []byte) error {
	var res GetUniversitiesOKBodyUniversitiesItems
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}