# pkg/models/post_login_o_k_body.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	strfmt "github.com/go-openapi/strfmt"

	"github.com/go-openapi/errors"
	"github.com/go-openapi/swag"
	"github.com/go-openapi/validate"
)

// PostLoginOKBody post login o k body
// swagger:model postLoginOKBody
type PostLoginOKBody struct {

	// token
	// Required: true
	Token string `json:"token"`
}

// Validate validates this post login o k body
func (m *PostLoginOKBody) Validate(formats strfmt.Registry) error {
	var res []error

	if err := m.validateToken(formats); err != nil {
		res = append(res, err)
	}

	if len(res) > 0 {
		return errors.CompositeValidationError(res...)
	}
	return nil
}

func (m *PostLoginOKBody) validateToken(formats strfmt.Registry) error {

	if err := validate.RequiredString("token", "body", string(m.Token)); err != nil {
		return err
	}

	return nil
}

// MarshalBinary interface implementation
func (m *PostLoginOKBody) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *PostLoginOKBody) UnmarshalBinary(b []byte) error {
	var res PostLoginOKBody
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}