# pkg/models/post_login_params_body.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	strfmt "github.com/go-openapi/strfmt"

	"github.com/go-openapi/errors"
	"github.com/go-openapi/swag"
	"github.com/go-openapi/validate"
)

// PostLoginParamsBody post login params body
// swagger:model postLoginParamsBody
type PostLoginParamsBody struct {

	// ключ администратора
	// Required: true
	APIKey string `json:"api_key"`
}

// Validate validates this post login params body
func (m *PostLoginParamsBody) Validate(formats strfmt.Registry) error {
	var res []error

	if err := m.validateAPIKey(formats); err != nil {
		res = append(res, err)
	}

	if len(res) > 0 {
		return errors.CompositeValidationError(res...)
	}
	return nil
}

func (m *PostLoginParamsBody) validateAPIKey(formats strfmt.Registry) error {

	if err := validate.RequiredString("api_key", "body", string(m.APIKey)); err != nil {
		return err
	}

	return nil
}

// MarshalBinary interface implementation
func (m *PostLoginParamsBody) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *PostLoginParamsBody) UnmarshalBinary(b []byte) error {
	var res PostLoginParamsBody
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}