# pkg/models/university_data_faculties_items.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	strfmt "github.com/go-openapi/strfmt"

	"github.com/go-openapi/errors"
	"github.com/go-openapi/swag"
	"github.com/go-openapi/validate"
)

// UniversityDataFacultiesItems university data faculties items
// swagger:model universityDataFacultiesItems
type UniversityDataFacultiesItems struct {

	// conditions
	// Required: true
	Conditions []string `json:"conditions"`

	// name
	// Required: true
	Name string `json:"name"`
}

// Validate validates this university data faculties items
func (m *UniversityDataFacultiesItems) Validate(formats strfmt.Registry) error {
	var res []error

	if err := m.validateConditions(formats); err != nil {
		res = append(res, err)
	}

	if err := m.validateName(formats); err != nil {
		res = append(res, err)
	}

	if len(res) > 0 {
		return errors.CompositeValidationError(res...)
	}
	return nil
}

func (m *UniversityDataFacultiesItems) validateConditions(formats strfmt.Registry) error {

	if err := validate.Required("conditions", "body", m.Conditions); err != nil {
		return err
	}

	return nil
}

func (m *UniversityDataFacultiesItems) validateName(formats strfmt.Registry) error {

	if err := validate.RequiredString("name", "body", string(m.Name)); err != nil {
		return err
	}

	return nil
}

// MarshalBinary interface implementation
func (m *UniversityDataFacultiesItems) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *UniversityDataFacultiesItems) UnmarshalBinary(b []byte) error {
	var res UniversityDataFacultiesItems
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}